Move `OrganizationList` component into org directory
